Got the progress bar working, as well as the `--test` command
