+++ .gitignore
@@ -1,4 +1,5 @@
 *~
 *.pyc
-*.*
+.*
+*.egg-info
 __pycache__
